Create separate Web Controller
