+++ Frontend/src/main/java/com/example/frontend/controllers/WebController.java
@@ -1,6 +1,10 @@
-package com.example.frontend;
+package com.example.frontend.controllers;
 
 
+import com.example.frontend.Image;
+import com.example.frontend.ImageClient;
+import com.example.frontend.ProductClient;
+import com.example.frontend.ProductDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -10,10 +14,11 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 
+// @RestController
 @AllArgsConstructor
 @Controller
 @RequestMapping("/")
-public class FrontendController {
+public class WebController {
     private final ImageClient imageClient;
     private final ProductClient productClient;
 
@@ -21,21 +26,6 @@ public class FrontendController {
     public String getHomePage(Model model) {
         //model.addAttribute("products", productClient.getAllProducts());
         return "index";
-    }
-
-    @GetMapping("images")
-    public List<Image> getImages(){
-        return imageClient.getImages();
-    }
-
-    @GetMapping("images/{id}")
-    public String getImageByName(@PathVariable Long id){
-        return imageClient.getImageByName(id);
-    }
-
-    @GetMapping("products")
-    public ResponseEntity<List<ProductDTO>> getAllProducts() {
-        return productClient.getAllProducts();
     }
 
 }
